requirements.txt
python-dotenv==1.0.1
pyjokes==0.8.3
pyTelegramBotAPI==4.23.0